fix: correct directory name from 'soba' to 'saba' in documentation and scripts; add initial project files for 'saba' including Cargo.toml, Makefile, and HTTP client implementation
