+++ AGENTS.md
@@ -2,17 +2,17 @@
 
 ## Build, Run & Dev Commands
 
-- Build app: `cd soba && make build`
+- Build app: `cd saba && make build`
   - Adds target `x86_64-unknown-none` and compiles release with required `RUSTFLAGS`.
-- Run under QEMU: `cd soba && ./run_on_wasabi.sh`
+- Run under QEMU: `cd saba && ./run_on_wasabi.sh`
   - Headless: `HEADLESS=1 ./run_on_wasabi.sh`
   - VNC viewer: `open vnc://localhost:5905`.
-- Lint/format: `cargo fmt` and `cargo clippy` (run in `soba/`).
+- Lint/format: `cargo fmt` and `cargo clippy` (run in `saba/`).
 - Prereqs (macOS): `brew install qemu wget jq`.
 
 ## Coding Style & Naming Conventions
 
-- Rust 2021; crates prefer `no_std` (OS-like env). Keep `main.rs` thin; move logic to `soba_core`.
+- Rust 2021; crates prefer `no_std` (OS-like env). Keep `main.rs` thin; move logic to `saba_core`.
 - Names: modules/files `snake_case`, types/enums `PascalCase`, functions/vars `snake_case`.
 - Formatting: `rustfmt` defaults. Avoid panics on hot paths; prefer `Result` and clear error messages.
 
@@ -25,5 +25,5 @@
   - “ownership/borrowing ≈ moving vs. referencing; think mutable vs. readonly refs.”
 - Explain `no_std`, `alloc`, and OS-specific macros (e.g., `entry_point!`) where used.
 - At module tops, describe purpose, inputs/outputs, and simple examples (one-liners or code blocks).
-- Favor examples over theory. Show tiny snippets: paths (`soba_core::url::parse`) and commands.
+- Favor examples over theory. Show tiny snippets: paths (`saba_core::url::parse`) and commands.
 - Keep comments updated when code changes; outdated comments must be removed or fixed.
