+++ saba/saba_core/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "soba_core"
+name = "saba_core"
 version = "0.1.0"
 edition = "2021"
