--- soba/soba_core/src/error.rs
@@ -1,9 +1,0 @@
-use alloc::string::String;
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Error {
-    Network(String),
-    UnexpectedInput(String),
-    InvalidUI(String),
-    Other(String),
-}
